Remove references to OpenMP
OpenMP is not used by this library. There is some code in the tests directory that isn't currently built that could use it.
Even so, it doesn't need to be linked to the main library or searched for.

Fix comparison of Fortran compiler to Clang

Remove set INTERFACE_LINK_LIBRARIES. There's no need to do that.